requirements.txt
python-dateutil
numpy
pandas
pytz
requests
pyzmq
pymongo
mongoengine
pip
wheel
cryptography
py
lxml
setuptools
TA-Lib
python-telegram-bot
paho-mqtt
git+https://github.com/imvinaypatil/webull.git@v1.0.1
backtrader
